django_files/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-countries==7.5
django-environ==0.9.0
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
typing_extensions==4.4.0
django-formtools==2.4